Compatibility with ARS 3.x (working around cURL) – PART II
When you send an array parameter to cURL for a POST request it gets squashed to the literal string "Array" instead of being sent as an array parameter. For example, if your POST parameters are:
`$postParams = ['foo' => [1,2,3,4]];`
cURL will send the POST body
`foo=Array`
instead of the expected
`foo[]=1&foo[]=2&foo[]=3&foo[]=4`
The only workaround is to send a comma separated list of values manually, i.e.
`foo=1,2,3,4`
and have server-side code convert them back to array representation. Or use GET, but then we will most likely hit the max GET string length.

This time I addressed the issue fully by adding the array-to-comma-separated-values-string code in the ArmArs class