Fix sftpcurl
If the directory didn't exist the sftpcurl adapter didn't try to create
it. Neither did cURL. No error was reported by cURL, though, and there
was no indication that the upload never worked.

Now we check if the remote directory exists. If not we try to
recursively create it. For each subdirectory in the path we have to
check if it already exists. If not, we create it. Otherwise we carry on
to the next part. This means that the first upload to an nonexistent
directory takes its sweet time, whereas subsequent uploads are lightning
fast. Better than a silent failure, though!